Fixing AdminToolsConsoleApp

+++ backend/AdminToolsConsoleApp/HelperMethods/Executor.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace AdminToolsConsoleApp.HelperMethods
@@ -19,7 +20,7 @@ namespace AdminToolsConsoleApp.HelperMethods
         public async Task ShowAllAsync()
         {
             var emails = await repository.GetAllEmailsAsync();
-            Console.WriteLine("\nEmail adrese za upozorenja:");
+            Console.WriteLine("\nAll email addresses:");
             foreach (var email in emails)
             {
                 Console.WriteLine($"- {email}");
@@ -28,19 +29,39 @@ namespace AdminToolsConsoleApp.HelperMethods
 
         public async Task AddAsync()
         {
-            Console.Write("Unesi novu email adresu: ");
+            Console.Write("Input email: ");
             string email = Console.ReadLine();
+            if (!Regex.IsMatch(email, @"^[^@\s]+@[^@\s]+\.com$"))
+            {
+                Console.WriteLine("Invalid email format. Email must be in the form something@something.com");
+                return;
+            }
+
+            bool exists = await repository.EmailExistsAsync(email);
+            if (exists)
+            {
+                Console.WriteLine("Email already exists.");
+                return;
+            }
+
             await repository.AddEmailAsync(email);
-            Console.WriteLine("Email dodat.");
+            Console.WriteLine("Email added.");
         }
 
         public async Task DeleteAsync()
         {
-            Console.Write("Unesi email koji zelis obrisati: ");
+            Console.Write("Input email for deletion: ");
             string email = Console.ReadLine();
 
+            if (!Regex.IsMatch(email, @"^[^@\s]+@[^@\s]+\.com$"))
+            {
+                Console.WriteLine("Invalid email format. Email must be in the form something@something.com");
+                return;
+            }
+
+
             bool deleted = await repository.DeleteEmailAsync(email);
-            Console.WriteLine(deleted ? "Email obrisan." : "Email nije pronadjen.");
+            Console.WriteLine(deleted ? "Email deleted." : "Email not found.");
         }
     }
 }
